Delete deprecated Text class && Move local property to shared place
