+++ example/lib/widgets/time_stamp_embed_widget.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_quill/flutter_quill.dart' hide Text;
+import 'package:flutter_quill/flutter_quill.dart';
 
 class TimeStampEmbed extends Embeddable {
   const TimeStampEmbed(
