+++ lib/extensions.dart
@@ -1,6 +1,6 @@
 library flutter_quill.extensions;
 
-export 'src/models/documents/nodes/leaf.dart' hide Text;
+export 'src/models/documents/nodes/leaf.dart';
 export 'src/models/rules/insert.dart';
 export 'src/utils/platform.dart';
 export 'src/utils/string.dart';
